controller: test metrics
In order to separate results from different tests results it is better
to not test with the global instances. APIs for instantiating the
counters multiple times with different names and then using them for
the controller get introduced in a backwards-compatible way.

The tests themselves represent the current state of the code. Whether
they make sense hasn't been checked.