controller: fix metrics after certain kind of errors
Certain situations (volume already exists, provisioner does not
support a raw block volume, an internal error for getting the claim
reference, inconsistent annotation (race or bug), provisioning
ignoring a claim) were counted as "volume provisioned successfully".

This was the result of an ambiguity between "no error" = "do not
requeue" and "no error" = "success". This is solved such that in all
of these situations a special error is returned which causes the
provisioning to be counted as "failed" in the metrics, but then that
special error is suppressed before returning to the code which decides
about requeuing.