controller: enable rescheduling of pods
For dynamic PV provisioning with node, provisioner may fail because node
is wrong. `selectedNode` must be removed to notify scheduler to schedule
again.

Original code by Yecheng. Rebased and extended by Patrick Ohly:

Rescheduling must not be done for operations that are still in
progress (ProvisionInBackground) or that still have a reasonable
chance of succeeding (ProvisioningNoChange). Provisioners must
explicitly ask for rescheduling by implementing the ProvisionerExt
interface and returning ProvisioningReschedule.