controller: fix module imports
The controller.test binary and the example still imported without the
/v5 suffix. This resulted in spurious
"sigs.k8s.io/sig-storage-lib-external-provisioner v4.1.0+incompatible
// indirect" go.mod entries in projects using the
lib (https://github.com/kubernetes-csi/external-provisioner/pull/405#discussion_r401286340).

The example needs to use the sources that it was checked out
with. Otherwise it is impossible to change the API and the example in
a single commit and/or there is a risk that breaking changes go
undetected because the example continues to build with the unmodified
lib.