Fix removal of PV protection finalizer
removeFinalizer() should not modify the input array of finalizers, because
that is directly PV object finalizers and thus it would modify the source PV.

We need the initial PV intact to compute correct JSON patch from the
initial PV to a PV with removed finalizers.

Allocate a separate slice instead.